api/bert_seq_tb/requirements.txt
Flask
pandas
torch==1.8.1
sentence-transformers==1.1.0
scikit-learn